You now have to choose an option. And it works with the prosses command :)
