+++ src/main/java/sdu/software/climatewars/Text/Parser.java
@@ -37,6 +37,7 @@ public class Parser
         return new Command(commands.getCommandWord(word1), word2);
     }
 
+
     public void showCommands()
     {
         commands.showAll();
